Adicionando renderização dos botões da pokedetails
